# Murmur128: use Go 1.9 bits.RotateLeft64 functions