--- a/murmur128.go
+++ b/murmur128.go
@@ -4,6 +4,7 @@ import (
 	//"encoding/binary"
 	"hash"
 	"unsafe"
+	"math/bits"
 )
 
 const (
@@ -68,20 +69,20 @@ func (d *digest128) bmix(p []byte) (tail []byte) {
 		k1, k2 := t[0], t[1]
 
 		k1 *= c1_128
-		k1 = (k1 << 31) | (k1 >> 33) // rotl64(k1, 31)
+		k1 = bits.RotateLeft64(k1, 31)
 		k1 *= c2_128
 		h1 ^= k1
 
-		h1 = (h1 << 27) | (h1 >> 37) // rotl64(h1, 27)
+		h1 = bits.RotateLeft64(h1, 27)
 		h1 += h2
 		h1 = h1*5 + 0x52dce729
 
 		k2 *= c2_128
-		k2 = (k2 << 33) | (k2 >> 31) // rotl64(k2, 33)
+		k2 = bits.RotateLeft64(k2, 33)
 		k2 *= c1_128
 		h2 ^= k2
 
-		h2 = (h2 << 31) | (h2 >> 33) // rotl64(h2, 31)
+		h2 = bits.RotateLeft64(h2, 31)
 		h2 += h1
 		h2 = h2*5 + 0x38495ab5
 	}
@@ -117,7 +118,7 @@ func (d *digest128) Sum128() (h1, h2 uint64) {
 		k2 ^= uint64(d.tail[8]) << 0
 
 		k2 *= c2_128
-		k2 = (k2 << 33) | (k2 >> 31) // rotl64(k2, 33)
+		k2 = bits.RotateLeft64(k2, 33)
 		k2 *= c1_128
 		h2 ^= k2
 
@@ -147,7 +148,7 @@ func (d *digest128) Sum128() (h1, h2 uint64) {
 	case 1:
 		k1 ^= uint64(d.tail[0]) << 0
 		k1 *= c1_128
-		k1 = (k1 << 31) | (k1 >> 33) // rotl64(k1, 31)
+		k1 = bits.RotateLeft64(k1, 31)
 		k1 *= c2_128
 		h1 ^= k1
 	}