@spaolacci 2019-03-17
This doesn't improve performance because initial hand written implementation was crafted to trigger Go compiler optimization.

Go 1.9 is old enough now. Having an explicit rotate left is however nice - if only for readability - so I the patch is still positive.

Thanks for proposing that, I'll update rotate 32 as well along the same lines.